Introduced AutoRegisterRoutes. Introduced BaseHttpAttribute.cs abstract class and derives from Attribute. Introduced HttpGetAttribute.cs and HttpPostAttribute.cs. These attributes indicates wherever method in controller is with get or post method and optional path of the method. (by default GET method is placed in the routeTable and path is "/{controllerName}/{controllerMethod.Name}") All static files in folders Views, wwRoot and ViewTests are included (embedded) in output directory of the project. There is no need to "copy always" every added file inside these folders.
Now if you want to register route you need to create controller named as main folder in views folder and csHtml file, named as method in the controller.
if you want custom path to be added in the route table just place HttpAttribute over the method in controller with path in it. Example [HttpGet("/a")]
or manually register it in the Configure method in startup.cs